requirements.txt
Flask
opencv-python
imutils
tensorflow
numpy
Pillow
